+++ js/Demo/demoLayers.js
@@ -6,6 +6,7 @@ addLayer("c", {
         name: "Candies", // This is optional, only used in a few places, If absent it just uses the layer id.
         symbol: "C", // This appears on the layer's node. Default is the id with the first letter capitalized
         position: 0, // Horizontal position within a row. By default it uses the layer id and sorts in alphabetical order
+        marked: "discord.png",
         startData() { return {
             unlocked: true,
 			points: new Decimal(0),
@@ -426,6 +427,7 @@ addLayer("f", {
     canReset() {
         return tmp[this.layer].baseAmount.gte(tmp[this.layer].nextAt)
     },
+
     // This is also non minimal, a Clickable!
     clickables: {